+++ tsconfig.json
@@ -22,6 +22,5 @@
 		"node_modules",
 		"./src/components/QuizList/FourProngQuiz.tsx",
 		"./src/pages/QuizList.tsx"
-		// "./src/pages/Calendar.tsx"
 	]
 }
